[Fix] UserInfo 컴포넌트화
