+++ user-data.json
@@ -30,7 +30,7 @@
       "date": "2002-06-07",
       "comment": "Cheers!",
       "image": "",
-      "checked": true
+      "checked": false
     },
     {
       "id": 5,
